+++ src/set.c
@@ -8,9 +8,9 @@
 
 struct entry {
   struct buffer *value;
-  size_t count;
-  /* TODO use unsigned instead of size_t in a lot of places */
-  unsigned metadata;
+  /* TODO use these types instead of size_t in a lot of places */
+  uint_fast32_t count;
+  uint_fast64_t metadata;
   struct entry *next;
 };
 
@@ -71,9 +71,9 @@ enum attribute {
   count,
 };
 
-static size_t get_attribute(enum attribute attribute,
-                            const struct buffer_view *value,
-                            const struct set *set) {
+static uint_fast64_t get_attribute(enum attribute attribute,
+                                   const struct buffer_view *value,
+                                   const struct set *set) {
   if (is_empty(set)) {
     return 0;
   }
@@ -84,12 +84,13 @@ static size_t get_attribute(enum attribute attribute,
   return 0;
 }
 
-size_t get_count(const struct buffer_view *value, const struct set *set) {
+uint_fast32_t get_count(const struct buffer_view *value,
+                        const struct set *set) {
   return get_attribute(count, value, set);
 }
 
-unsigned get_last_metadata(const struct buffer_view *value,
-                           const struct set *set) {
+uint_fast64_t get_last_metadata(const struct buffer_view *value,
+                                const struct set *set) {
   return get_attribute(metadata, value, set);
 }
 
@@ -97,8 +98,8 @@ bool is_within(const struct buffer_view *value, const struct set *set) {
   return get_count(value, set);
 }
 
-void add_with_metadata(const char *value, unsigned metadata, struct set *set,
-                       struct trace *trace) {
+void add_with_metadata(const char *value, uint_fast64_t metadata,
+                       struct set *set, struct trace *trace) {
   struct buffer *buffer = create_buffer(trace);
   concat_string(value, buffer, trace);
   if (!ok(trace)) {
